+++ src/codegen/mod.rs
@@ -5,3 +5,5 @@ pub mod expr;
 pub use core::Codegen;
 
 mod landline;
+
+mod library;